+++ rdb/re_json.go
@@ -3,15 +3,17 @@ package rdb
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/go-redis/redis"
 	"strconv"
+
+	"github.com/go-redis/redis"
 )
 
 func JsonGet[T any](db *redis.Client, key, path string, container *T) error {
 	cmd := redis.NewStringCmd("JSON.GET", key, path)
+
 	db.Process(cmd)
-	fmt.Println(cmd)
 	v, err := cmd.Result()
+
 	_ = json.Unmarshal([]byte(v), &container)
 	return err
 }
@@ -36,6 +38,7 @@ func JsonArrAppend[T any](db *redis.Client, key, path string, container *T) erro
 
 func JsonArrLen(db *redis.Client, key, path string) (int, error) {
 	cmd := redis.NewSliceCmd("JSON.ARRLEN", key, path)
+
 	db.Process(cmd)
 	l, err := cmd.Result()
 	if err != nil {
@@ -43,4 +46,12 @@ func JsonArrLen(db *redis.Client, key, path string) (int, error) {
 	}
 	intL, err := strconv.Atoi(fmt.Sprintf("%v", l[0]))
 	return intL, err
+}
+
+func JsonDel(db *redis.Client, key, path string) error {
+	cmd := redis.NewSliceCmd("JSON.DEL", key, path)
+
+	db.Process(cmd)
+	err := cmd.Err()
+	return err
 }